chore: Add code to handle hello_world kind & Move all macros inside get_macros_code method to separate module inside composer-boilerplate
